refactor(core): Move `ApplicationError` to `@n8n/errors`
