+++ packages/@n8n/errors/src/types.ts
