+++ packages/workflow/src/errors/index.ts
@@ -1,9 +1,8 @@
-export type * from './error.types';
 export { BaseError, type BaseErrorOptions } from './base/base.error';
 export { OperationalError, type OperationalErrorOptions } from './base/operational.error';
 export { UnexpectedError, type UnexpectedErrorOptions } from './base/unexpected.error';
 export { UserError, type UserErrorOptions } from './base/user.error';
-export { ApplicationError } from './application.error';
+export { ApplicationError } from '@n8n/errors';
 export { ExpressionError } from './expression.error';
 export { ExecutionCancelledError } from './execution-cancelled.error';
 export { NodeApiError } from './node-api.error';
