+++ packages/workflow/src/errors/node-api.error.ts
@@ -4,7 +4,7 @@ import type { AxiosError } from 'axios';
 import { parseString } from 'xml2js';
 
 import { NodeError } from './abstract/node.error';
-import type { ErrorLevel } from './error.types';
+import type { ErrorLevel } from '@n8n/errors';
 import {
 	NO_OP_NODE_TYPE,
 	UNKNOWN_ERROR_DESCRIPTION,
