+++ packages/workflow/src/errors/node-operation.error.ts
@@ -1,5 +1,5 @@
 import { NodeError } from './abstract/node.error';
-import { ApplicationError } from './application.error';
+import { ApplicationError } from '@n8n/errors';
 import type { NodeOperationErrorOptions } from './node-api.error';
 import type { INode, JsonObject } from '../interfaces';
