+++ packages/workflow/src/errors/trigger-close.error.ts
@@ -1,5 +1,4 @@
-import { ApplicationError } from './application.error';
-import type { ErrorLevel } from './error.types';
+import { ApplicationError, type ErrorLevel } from '@n8n/errors';
 import type { INode } from '../interfaces';
 
 interface TriggerCloseErrorOptions extends ErrorOptions {
