+++ packages/workflow/src/expression.ts
@@ -1,6 +1,6 @@
 import { DateTime, Duration, Interval } from 'luxon';
 
-import { ApplicationError } from './errors/application.error';
+import { ApplicationError } from '@n8n/errors';
 import { ExpressionExtensionError } from './errors/expression-extension.error';
 import { ExpressionError } from './errors/expression.error';
 import { evaluateExpression, setErrorHandler } from './expression-evaluator-proxy';
