+++ packages/workflow/src/workflow-data-proxy.ts
@@ -7,7 +7,7 @@ import { DateTime, Duration, Interval, Settings } from 'luxon';
 
 import { augmentArray, augmentObject } from './augment-object';
 import { AGENT_LANGCHAIN_NODE_TYPE, SCRIPTING_NODE_TYPES } from './constants';
-import { ApplicationError } from './errors/application.error';
+import { ApplicationError } from '@n8n/errors';
 import {
 	ExpressionError,
 	type ExpressionErrorOptions,
